package-json-version normalization strategy chagned.
The normalization of excess dots in package-json-version input is changed to remove excess dots instead of replacing them with dashes (-).

Believe this is a better strategy for our current use, to get a patch version that has a single number before the first dash, which corresponds to the date and time it was built.

This should display the whole timestamp when doing yarn interactive-upgrade, instead of just the year that was displayed with the old strategy of replacing dots with dashes.