Removal of pre-release part of package version.
Because the patch version part passed to the clientmaker from the build pipeline, and used to generate the javascript package version,  for our usecase contains a dash (-), it was interpreted as being a pre release. This caused the packages published with the tool to not be picked up by dependabot.

With this change we start removing such pre-release parts of the patch version string, so that published packages are interpreted as GA versions and picked up by dependabot in the repositories depending on it.