Change to use new @hey-api/openapi-ts package as code generator.
The @hey-api/openapi-ts package is the successor the original tool, and is where development is happening.

This transition keeps the generated output compatible with previous output. Based on testing, no changes need to be made in the projects using generated output.

But it required changes to input openapi.json. Previous code generator created valid typescript from some enums that were printed with values such as "ENUM_NAME('VALUE')". It did so by escaping the ' character within the parenthesis. This new version did not do such escaping. This caused generated code to fail typescript compilation.

This is ok since such enums should not be part of the api. Their toString() method should instead be fixed to return the base enum value which the code generator creates valid typescript from.