Upgrade to hey-api/openapi-ts version 0.60.1. Breaking change!
This upgrade changes to generate flat function based client instead of old class based. It adds a custom methodNameBuilder that prefixes the generatedrequest functions with tag_, to make it easier to find the correct function since the old lookup by class property is gone.

It also uses new experimental parser, and bundles the fetch based client.

This new version does not generate separate enum consts/types if there is a name collision, so for this to work well, steps should be taken to avoid having enums with same name.

Recommended steps are to make the openapi.json file have enums as refs instead of inline.

These changes make this generator upgrade very breaking, so will update version to 2.0.0