Configure code generator to preserve case of typenames generated. (#19)
* Configure code generator to preserve case of typenames generated.

This makes it so that when we create client code from openapi spec that has fully qualified type names, we get a snake case pattern for the package name part, and PascalCase for the class name part. This makes it a lot easier to read the generated typescript name, and see what is the package name and what is the class name.

* built javascript in dist/ updated by yarn compile

---------

Co-authored-by: GitHub Action <action@github.com>